% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutations_func.R
\name{get_ShuffledPositions_vFunct}
\alias{get_ShuffledPositions_vFunct}
\title{Gene-to-TAD permutations}
\usage{
get_ShuffledPositions_vFunct(g2TADdt, geneIDlist, nClass, TADonly,
  withExprClass, geneAggregExpressionDT = NULL, rd_idx = 0)
}
\arguments{
\item{g2TADdt}{Gene-to-TAD dataframe (expected columns: entrezID, region).}

\item{geneIDlist}{List of gene IDs that should be used.}

\item{nClass}{The number of classes of expression in which genes are shuffled.}

\item{TADonly}{If only genes from TADs should be used.}

\item{withExprClass}{If shuffling should take place within classes.}

\item{geneAggregExpressionDT}{Dataframe with aggregated gene expression (across samples).}

\item{rd_idx}{For setting seed.}
}
\value{
The permtuation data.
}
\description{
Performs the permutation (for 1 permutation).
}
