% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_func.R
\name{noZero_breaks}
\alias{noZero_breaks}
\title{Adapt pretty breaks with no zeros}
\usage{
noZero_breaks(n = 5, ...)
}
\arguments{
\item{n}{The number of breaks wanted.}

\item{...}{Other parameters passed to scales::force_all.}
}
\value{
The breaks without zeros
}
\description{
Adaptation to scales::pretty breaks to remove the zeros of the scale.
}
